+++ src/main/php/util/URIParameters.class.php
@@ -16,14 +16,24 @@ class URIParameters implements Value, \IteratorAggregate {
    * Creates an instance from either a URI or a string
    *
    * @param  util.URI|string $arg
+   */
+  public function __construct($arg) {
+    $this->pairs= self::decode($arg instanceof URI ? $arg->query() : $arg);
+  }
+
+  /**
+   * Creates an instance from either a URI or a string
+   *
+   * @param  string $query
    * @param  int $nesting Maximum nesting depth for arrays
    * @see    http://php.net/manual/en/info.configuration.php#ini.max-input-nesting-level
+   * @return [:var]
    */
-  public function __construct($arg, $nesting= 0) {
-    $query= $arg instanceof URI ? $arg->query() : $arg;
-    if ('' === $query) return;
+  public static function decode($query, $nesting= 0) {
+    if ('' === $query) return [];
 
     $nesting || $nesting= ini_get('max_input_nesting_level') ?: 64;
+    $pairs= [];
     foreach (explode('&', $query) as $pair) {
       $key= $value= null;
       sscanf($pair, "%[^=]=%[^\r]", $key, $value);
@@ -36,10 +46,10 @@ class URIParameters implements Value, \IteratorAggregate {
 
       if ($start= strpos($key, '[')) {    // Array notation
         $base= substr($key, 0, $start);
-        if (!isset($this->pairs[$base]) || !is_array($this->pairs[$base])) {
-          $this->pairs[$base]= [];
+        if (!isset($pairs[$base]) || !is_array($pairs[$base])) {
+          $pairs[$base]= [];
         }
-        $ptr= &$this->pairs[$base];
+        $ptr= &$pairs[$base];
         $offset= 0;
         $level= 0;
         do {
@@ -57,9 +67,10 @@ class URIParameters implements Value, \IteratorAggregate {
         } while ($start= strpos($key, '[', $offset));
         $ptr= rawurldecode($value);
       } else {
-        $this->pairs[$key]= rawurldecode($value);
+        $pairs[$key]= rawurldecode($value);
       }
     }
+    return $pairs;
   }
 
   /** @return int */
@@ -94,7 +105,7 @@ class URIParameters implements Value, \IteratorAggregate {
    * @param  string $offset
    * @return string
    */
-  private function pair($key, $value, $offset= '') {
+  private static function pair($key, $value, $offset= '') {
     $query= '';
     if ('' === $value) {
       $query.= '&'.rawurlencode($key).$offset;
@@ -102,11 +113,11 @@ class URIParameters implements Value, \IteratorAggregate {
       if (0 === key($value)) {
         $offset.= '[]';
         foreach ($value as $v) {
-          $query.= $this->pair($key, $v, $offset);
+          $query.= self::pair($key, $v, $offset);
         }
       } else {
         foreach ($value as $k => $v) {
-          $query.= $this->pair($key, $v, $offset.'['.rawurlencode($k).']');
+          $query.= self::pair($key, $v, $offset.'['.rawurlencode($k).']');
         }
       }
     } else {
@@ -115,15 +126,19 @@ class URIParameters implements Value, \IteratorAggregate {
     return $query;
   }
 
-  /** @return string */
-  public function __toString() {
-    if (empty($this->pairs)) return '';
+  public static function encode($pairs) {
+    if (empty($pairs)) return '';
 
     $query= '';
-    foreach ($this->pairs as $key => $value) {
-      $query.= $this->pair($key, $value);
+    foreach ($pairs as $key => $value) {
+      $query.= self::pair($key, $value);
     }
     return substr($query, 1);
+  }
+
+  /** @return string */
+  public function __toString() {
+    return self::encode($this->pairs);
   }
 
   /** @return string */